+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### BREAKING CHANGES
 
+- `\Wizaplace\Favorite\FavoriteService::isInFavorites` is now taking a `string $declinationId` parameter instead of a `int $declinationId` one.
+- `\Wizaplace\Favorite\FavoriteService::addDeclinationToUserFavorites` is now taking a `string $declinationId` parameter instead of a `int $declinationId` one.
+
 ### New features
 
 - Added `\Wizaplace\Basket\BasketItem::getDeclinationOptions`